requirements.txt
telethon
asyncio
tkintertable
pysqlite3
threaded
pycryptodome
